Add tweaks to thread + add sdl_net

+++ SDL-1.2.15/src/video/switch/SDL_switchvideo.h
@@ -29,6 +29,7 @@
 /* Hidden "this" pointer for the video functions */
 #define _THIS	SDL_VideoDevice *this
 
+//typedef unsigned u8;
 
 /* Private display data */
 
@@ -38,7 +39,7 @@ struct SDL_PrivateVideoData {
     int bpp;
     void *buffer;
 	unsigned int flags; // backup of create device flags
-	u8 clearbuffer; // number of full hw buffer to clear at startup and after toggling fullscreen (and maybe other situations in future)
+	unsigned clearbuffer; // number of full hw buffer to clear at startup and after toggling fullscreen (and maybe other situations in future)
 };
 
 #endif /* _SDL_nullvideo_h */
